[SofaDefaultType] Add StdRigidTypes in fwd.h
